Move view examples to example_test.go
